+++ agents/builder/src/server/routes/submission.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify
-from src.server.services import database
+from src.database import get_db, Submission
 import logging
 
 logger = logging.getLogger(__name__)
@@ -11,27 +11,24 @@ bp = Blueprint("submission", __name__)
 def fetch_submission(roundNumber):
     logger.info(f"Fetching submission for round: {roundNumber}")
 
-    db = database.get_db()
-    cursor = db.cursor()
-    query = cursor.execute(
-        """
-        SELECT roundNumber, status, pr_url, username, repo_owner, repo_name
-        FROM submissions
-        WHERE roundNumber = ? and status = 'completed'
-        """,
-        (int(roundNumber),),
+    db = get_db()
+    submission = (
+        db.query(Submission)
+        .filter(
+            Submission.round_number == int(roundNumber),
+            Submission.status == "completed",
+        )
+        .first()
     )
-    result = query.fetchone()
-    database.close_db()
 
-    if result:
+    if submission:
         return jsonify(
             {
-                "roundNumber": result["roundNumber"],
-                "prUrl": result["pr_url"],
-                "githubUsername": result["username"],
-                "repoOwner": result["repo_owner"],
-                "repoName": result["repo_name"],
+                "roundNumber": submission.round_number,
+                "prUrl": submission.pr_url,
+                "githubUsername": submission.username,
+                "repoOwner": submission.repo_owner,
+                "repoName": submission.repo_name,
             }
         )
     else:
