--- agents/builder/src/server_new/routes/submit_pr.py
@@ -1,20 +1,0 @@
-from flask import Blueprint, jsonify, request
-from src.server_new.services.task_service import submit_pr
-
-bp = Blueprint("submit_pr", __name__)
-
-
-@bp.route("/submit-pr/<roundNumber>", methods=["POST"])
-def submit_pr_route(roundNumber):
-    data = request.get_json()
-    signature = data.get("signature")
-    staking_key = data.get("stakingKey")
-    pub_key = data.get("pubKey")
-    pr_url = data.get("prUrl")
-
-    if not pr_url:
-        return jsonify({"error": "Missing PR URL"}), 400
-
-    message = submit_pr(signature, staking_key, pub_key, pr_url, roundNumber)
-
-    return jsonify({"message": message})
